Refactor: move common functions to spdx_common.py
This commit moves common functions from various helper files in the
spdx/spdxtagvalue subdirectory to the spdx/spdx_common.py file. This
commit also updates the helper files in the spdx/spdxtagvalue directory
to reference the new location of the helper functions in spdx_common.py.
The helper files that are now in spdx_common.py will be used by both the
tag-value and JSON SPDX formats for document creation.

Works towards #839
