+++ tern/formats/spdx/spdxtagvalue/layer_helpers.py
@@ -7,24 +7,17 @@
 Helpers for layer information
 Layers for SPDX act like a Package
 """
-import hashlib
 import logging
 import os
 
+from tern.formats.spdx import spdx_common
 from tern.formats.spdx.spdxtagvalue import formats as spdx_formats
 from tern.utils import constants
 from tern.formats.spdx.spdxtagvalue import file_helpers as fhelpers
-from tern.formats.spdx.spdxtagvalue import package_helpers as phelpers
 from tern.report import content
 
 # global logger
 logger = logging.getLogger(constants.logger_name)
-
-
-def get_layer_spdxref(layer_obj):
-    '''Given the layer object, return an SPDX reference ID'''
-    # here we return the shortened diff_id of the layer
-    return 'SPDXRef-{}'.format(layer_obj.diff_id[:10])
 
 
 def get_layer_comment(layer_obj):
@@ -48,11 +41,11 @@ def get_layer_package_relationships(layer_obj):
     objects to packages. This is usually of the form:
         layer SPDXIP CONTAINS package SPDXID'''
     block = ''
-    layer_reference = get_layer_spdxref(layer_obj)
+    layer_reference = spdx_common.get_layer_spdxref(layer_obj)
     for package in layer_obj.packages:
         block = block + spdx_formats.contains.format(
             outer=layer_reference,
-            inner=phelpers.get_package_spdxref(package)) + '\n'
+            inner=spdx_common.get_package_spdxref(package)) + '\n'
     return block
 
 
@@ -61,50 +54,8 @@ def get_layer_prereq(curr_layer, prev_layer):
     between the current and previous layer. This should look like this:
         curr_layer HAS_PREREQUISITE prev_layer'''
     return spdx_formats.prereq.format(
-        after=get_layer_spdxref(curr_layer),
-        before=get_layer_spdxref(prev_layer))
-
-
-def get_layer_verification_code(layer_obj):
-    '''Calculate the verification code from the files in an image layer. This
-    assumes that layer_obj.files_analyzed is True. The implementation follows
-    the algorithm in the SPDX spec v 2.1 which requires SHA1 to be used to
-    calculate the checksums of the file and the final verification code'''
-    sha1_list = []
-    for filedata in layer_obj.files:
-        filesha = filedata.get_checksum('sha1')
-        if not filesha:
-            # we cannot create a verification code, hence file generation
-            # is aborted
-            logger.critical(
-                'File %s does not have a sha1 checksum. Failed to generate '
-                'a SPDX tag-value report', filedata.path)
-            return None
-        sha1_list.append(filesha)
-    sha1_list.sort()
-    sha1s = ''.join(sha1_list)
-    return hashlib.sha1(sha1s.encode('utf-8')).hexdigest()  # nosec
-
-
-def get_layer_checksum(layer_obj):
-    '''Return a SPDX formatted checksum value. It should be of the form:
-        checksum_type: <checksum>'''
-    return '{}: {}'.format(layer_obj.checksum_type.upper(), layer_obj.checksum)
-
-
-def get_layer_licenses(layer_obj):
-    '''Return a list of unique licenses from the files analyzed
-    in the layer object. It is assumed that the files were analyzed and
-    there should be some license expressions. If there are not, an empty list
-    is returned'''
-    licenses = set()
-    for filedata in layer_obj.files:
-        # we will use the SPDX license expressions here as they will be
-        # valid SPDX license identifiers
-        if filedata.licenses:
-            for lic in fhelpers.get_file_licenses(filedata):
-                licenses.add(lic)
-    return list(licenses)
+        after=spdx_common.get_layer_spdxref(curr_layer),
+        before=spdx_common.get_layer_spdxref(prev_layer))
 
 
 def get_package_license_info_block(layer_obj):
@@ -113,7 +64,7 @@ def get_package_license_info_block(layer_obj):
     not analyzed'''
     block = ''
     if layer_obj.files_analyzed:
-        licenses = get_layer_licenses(layer_obj)
+        licenses = spdx_common.get_layer_licenses(layer_obj)
         if licenses:
             for lic in licenses:
                 block += 'PackageLicenseInfoFromFiles: {}\n'.format(
@@ -129,7 +80,7 @@ def get_layer_file_data_block(layer_obj, template):
     analyzed'''
     block = ''
     if layer_obj.files_analyzed:
-        layer_checksum = get_layer_checksum(layer_obj)
+        layer_checksum = spdx_common.get_layer_checksum(layer_obj)
         # insert a blank line in the beginning
         block += '\n'
         # some files are located in different places in the filesystem
@@ -139,7 +90,7 @@ def get_layer_file_data_block(layer_obj, template):
         # file data
         for filedata in layer_obj.files:
             # we use the layer checksum as the layer id
-            file_ref = fhelpers.get_file_spdxref(filedata, layer_checksum)
+            file_ref = spdx_common.get_file_spdxref(filedata, layer_checksum)
             if file_ref not in file_refs:
                 block += fhelpers.get_file_block(
                     filedata, template, layer_checksum) + '\n'
@@ -163,7 +114,7 @@ def get_layer_block(layer_obj, template, image_loc=''):
     # Package Name
     block += 'PackageName: {}\n'.format(os.path.basename(layer_obj.tar_file))
     # Package SPDXID
-    block += 'SPDXID: {}\n'.format(get_layer_spdxref(layer_obj))
+    block += 'SPDXID: {}\n'.format(spdx_common.get_layer_spdxref(layer_obj))
     # Package File Name
     block += 'PackageFileName: {}\n'.format(layer_obj.tar_file)
     # Package Download Location
@@ -177,11 +128,12 @@ def get_layer_block(layer_obj, template, image_loc=''):
         # we need a package verification code
         block += 'FilesAnalyzed: true\n'
         block += 'PackageVerificationCode: {}\n'.format(
-            get_layer_verification_code(layer_obj))
+            spdx_common.get_layer_verification_code(layer_obj))
     else:
         block += 'FilesAnalyzed: false\n'
     # Package Checksum
-    block += 'PackageChecksum: {}\n'.format(get_layer_checksum(layer_obj))
+    block += 'PackageChecksum: {}\n'.format(
+        spdx_common.get_layer_checksum(layer_obj))
     # Package License Concluded (always NOASSERTION)
     block += 'PackageLicenseConcluded: NOASSERTION\n'
     # All licenses info from files if files_analyzed is true
